NumpadView and DiskView refactoring. Using a common class.
